requirements.txt
flask
autopep8
python-dotenv
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
email_validator
flask-socketio
pika
requests
pytest